fix for copying of vm log
vm log was tried to be copied for non VM jobs, too, where it doesn't exist

+++ src/common/root-functions.sh
@@ -486,6 +486,8 @@ getVMCountOnLocalhost() {
 # copy libvirt log from global logs dir to vm job dir (for vm jobs)
 #
 copyVMlogFile() {
-  cp /var/log/libvirt/qemu/${JOBID}*.log "$VM_JOB_DIR/$LOCALHOST/";
-  chown $USERNAME:$USERNAME "$VM_JOB_DIR/$LOCALHOST/${JOBID}*.log";
+  if [ -n "$(ls /var/log/libvirt/qemu/ | grep $JOBID | grep -E \.log$)" ]; then
+    cp /var/log/libvirt/qemu/${JOBID}*.log "$VM_JOB_DIR/$LOCALHOST/";
+    chown $USERNAME:$USERNAME "$VM_JOB_DIR/$LOCALHOST/${JOBID}*.log";
+  fi
 }
